# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/forms": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the list of forms you created",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master - Forms"
                ],
                "summary": "List Forms",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page of list data",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limitting data you want to get",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Find your data with keywords",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "created_at:desc",
                        "description": "Ordering data",
                        "name": "orderBy",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request success",
                        "schema": {
                            "$ref": "#/definitions/commonschema.ResponseHTTP"
                        }
                    },
                    "400": {
                        "description": "Request failure",
                        "schema": {
                            "$ref": "#/definitions/commonschema.ResponseHTTP"
                        }
                    }
                }
            }
        },
        "/api/forms/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detail data of form you choose",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master - Forms"
                ],
                "summary": "Detail Form",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of your data",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request success",
                        "schema": {
                            "$ref": "#/definitions/commonschema.ResponseHTTP"
                        }
                    },
                    "400": {
                        "description": "Request failure",
                        "schema": {
                            "$ref": "#/definitions/commonschema.ResponseHTTP"
                        }
                    }
                }
            }
        },
...
        "/api/workspaces": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the list of workspaces you created",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master - Workspaces"
                ],
                "summary": "List Workspaces",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page of list data",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Limitting data you want to get",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Find your data with keywords",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "created_at:desc",
                        "description": "Ordering data",
                        "name": "orderBy",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request success",
                        "schema": {
                            "$ref": "#/definitions/commonschema.ResponseHTTP"
                        }
                    },
                    "400": {
                        "description": "Request failure",
                        "schema": {
                            "$ref": "#/definitions/commonschema.ResponseHTTP"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create new workspace data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master - Workspaces"
                ],
                "summary": "Create Workspace",
                "parameters": [
                    {
                        "description": "Workspace payload",
                        "name": "workspacePayload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/masterschema.WorkspacePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request success",
                        "schema": {
                            "$ref": "#/definitions/commonschema.ResponseHTTP"
                        }
                    },
                    "400": {
                        "description": "Request failure",
                        "schema": {
                            "$ref": "#/definitions/commonschema.ResponseHTTP"
                        }
                    }
                }
            }
        },
        "/api/workspaces/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detail data of workspace you choose",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master - Workspaces"
                ],
                "summary": "Detail Workspace",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of your data",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request success",
                        "schema": {
                            "$ref": "#/definitions/commonschema.ResponseHTTP"
                        }
                    },
                    "400": {
                        "description": "Request failure",
                        "schema": {
                            "$ref": "#/definitions/commonschema.ResponseHTTP"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update existing workspace data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master - Workspaces"
                ],
                "summary": "Update Workspace",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of your data",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Workspace payload",
                        "name": "workspacePayload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/masterschema.WorkspacePayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request success",
                        "schema": {
                            "$ref": "#/definitions/commonschema.ResponseHTTP"
                        }
                    },
                    "400": {
                        "description": "Request failure",
                        "schema": {
                            "$ref": "#/definitions/commonschema.ResponseHTTP"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete existing workspace data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Master - Workspaces"
                ],
                "summary": "Delete Workspace",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID of your data",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request success",
                        "schema": {
                            "$ref": "#/definitions/commonschema.ResponseHTTP"
                        }
                    },
                    "400": {
                        "description": "Request failure",
                        "schema": {
                            "$ref": "#/definitions/commonschema.ResponseHTTP"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "commonschema.ResponseHTTP": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "masterschema.WorkspacePayload": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "is_publish": {
                    "type": "boolean",
                    "default": false
                },
                "thumbnail": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}